adding new command to generate docs for all plugins
